backend/requirements.txt
anyio==3.6.1
arrow==1.2.2
asgiref==3.5.2
bcrypt==3.2.2
binaryornot==0.4.4
certifi==2022.5.18.1
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cookiecutter==1.7.3
cryptography==37.0.2
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
Jinja2==3.1.2
jinja2-time==0.2.0
jose==1.0.0
MarkupSafe==2.1.1
passlib==1.7.4
poyo==0.5.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
python-slugify==6.1.2
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.19.1
text-unidecode==1.3
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
